+++ src/main/java/com/robosh/model/entities/Coupon.java
@@ -20,7 +20,7 @@ import javax.persistence.Id;
 public class Coupon {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private Integer idCoupon;
+    private Long idCoupon;
     private String couponName;
     private int discount;
 }
