+++ src/main/java/com/robosh/repository/AddressRepository.java
@@ -3,6 +3,6 @@ package com.robosh.repository;
 import com.robosh.model.entities.Address;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AddressRepository extends JpaRepository<Address,Integer> {
-    Address findByIdAddress(Integer id);
+public interface AddressRepository extends JpaRepository<Address,Long> {
+    Address findByIdAddress(Long id);
 }
